+++ builtins/src/nodebox/builtins/draw/Rect.java
@@ -18,7 +18,7 @@ public class Rect extends DrawingNode {
     public final FloatPort pStrokeWeight = new FloatPort(this, "strokeWeight", Port.Direction.INPUT, 1f);
 
     @Override
-    public void draw(PApplet g, Context context, double time) {
+    public void draw(PApplet g, Context context, float time) {
         ProcessingSupport.setStyle(g, pFill, pStroke, pStrokeWeight);
         g.rect(pX.get(), pY.get(), pWidth.get(), pHeight.get());
     }
